backend/requirements.txt
Django==6.0.3
djangorestframework==3.14.0
djangorestframework-simplejwt>=5.4.0
django-allauth>=0.63.0
django-cors-headers==4.3.1
psycopg2-binary>=2.9.0
python-decouple==3.8
Pillow>=11.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
gunicorn>=21.0.0
whitenoise>=6.0.0
twilio>=9.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
cryptography>=38.0.3
